+++ inference_models/inference_models/models/optimization/__init__.py
@@ -1,0 +1,1 @@
+"""Reusable contracts and selection utilities for optimized inference paths."""
