+++ inference_models/inference_models/models/optimization/ids.py
@@ -1,0 +1,4 @@
+"""Stable conventional IDs shared by inference-path optimizations."""
+
+BASE_IMPLEMENTATION_ID = "base"
+AUTO_IMPLEMENTATION_ID = "auto"
