+++ inference_models/inference_models/models/optimization/registry.py
@@ -1,4 +1,4 @@
-"""Context-aware registry for RF-DETR stage implementations."""
+"""Context-aware registry for selectable inference-stage implementations."""
 
 from __future__ import annotations
 
@@ -6,17 +6,30 @@ from collections import defaultdict
 from typing import DefaultDict, Dict, Iterable, Tuple
 
 from inference_models.errors import ModelRuntimeError
-from inference_models.models.rfdetr.optimization.contracts import (
+from inference_models.models.optimization.contracts import (
     ExecutionContext,
     InferenceStage,
     OptimizationStage,
 )
+from inference_models.models.optimization.ids import (
+    AUTO_IMPLEMENTATION_ID,
+    BASE_IMPLEMENTATION_ID,
+)
 
 
 class ImplementationRegistry:
-    """Register and resolve typed RF-DETR stage implementations."""
+    """Register and resolve typed inference-stage implementations."""
 
-    def __init__(self) -> None:
+    def __init__(
+        self,
+        *,
+        scope_name: str,
+        base_id: str = BASE_IMPLEMENTATION_ID,
+        auto_id: str = AUTO_IMPLEMENTATION_ID,
+    ) -> None:
+        self._scope_name = scope_name
+        self._base_id = base_id
+        self._auto_id = auto_id
         self._implementations: DefaultDict[
             OptimizationStage, Dict[str, InferenceStage]
         ] = defaultdict(dict)
@@ -50,7 +63,7 @@ class ImplementationRegistry:
 
         Args:
             stage: Stage category being selected.
-            requested_id: Stable implementation ID or ``auto``.
+            requested_id: Stable implementation ID or the automatic-selection ID.
             context: Runtime target and request context.
 
         Returns:
@@ -60,20 +73,22 @@ class ImplementationRegistry:
             ModelRuntimeError: If the ID is unknown or incompatible.
         """
         stage_implementations = self._implementations.get(stage, {})
-        if requested_id == "auto":
-            return self._resolve_auto(
+        if requested_id == self._auto_id:
+            implementation = self._resolve_auto(
                 stage=stage,
                 implementations=stage_implementations.values(),
                 context=context,
             )
 
+            return implementation
+
         implementation = stage_implementations.get(requested_id)
         if implementation is None:
-            available = ", ".join(sorted(["auto", *stage_implementations]))
+            available = ", ".join(sorted([self._auto_id, *stage_implementations]))
             raise ModelRuntimeError(
                 message=(
-                    f"Unknown RF-DETR {stage.value} implementation {requested_id!r}. "
-                    f"Available implementations: {available}."
+                    f"Unknown {self._scope_name} {stage.value} implementation "
+                    f"{requested_id!r}. Available implementations: {available}."
                 ),
                 help_url=(
                     "https://inference-models.roboflow.com/errors/models-runtime/"
@@ -83,8 +98,9 @@ class ImplementationRegistry:
         if not implementation.is_compatible(context):
             raise ModelRuntimeError(
                 message=(
-                    f"RF-DETR {stage.value} implementation {requested_id!r} is not "
-                    f"compatible with device={context.device!r}, "
+                    f"{self._scope_name} {stage.value} implementation "
+                    f"{requested_id!r} is not compatible with "
+                    f"device={context.device!r}, "
                     f"device_kind={context.device_kind!r}, "
                     f"device_family={context.device_family!r}."
                 ),
@@ -122,7 +138,7 @@ class ImplementationRegistry:
         implementations = tuple(implementations)
         for implementation in implementations:
             metadata = implementation.metadata
-            if metadata.implementation_id == "base":
+            if metadata.implementation_id == self._base_id:
                 continue
             validated = any(
                 environment.matches(context)
@@ -135,13 +151,16 @@ class ImplementationRegistry:
             (
                 implementation
                 for implementation in implementations
-                if implementation.metadata.implementation_id == "base"
+                if implementation.metadata.implementation_id == self._base_id
             ),
             None,
         )
         if base is None:
             raise ModelRuntimeError(
-                message=f"RF-DETR {stage.value} registry has no base implementation.",
+                message=(
+                    f"{self._scope_name} {stage.value} registry has no "
+                    f"{self._base_id!r} implementation."
+                ),
                 help_url=(
                     "https://inference-models.roboflow.com/errors/models-runtime/"
                     "#modelruntimeerror"
