+++ inference_models/inference_models/models/rfdetr/optimization/ids.py
@@ -1,7 +1,8 @@
 """Stable IDs for RF-DETR inference-path implementations."""
 
-RFDETR_PREPROCESSOR_BASE = "base"
-RFDETR_PREPROCESSOR_AUTO = "auto"
+from inference_models.models.optimization.ids import BASE_IMPLEMENTATION_ID
+
+RFDETR_PREPROCESSOR_BASE = BASE_IMPLEMENTATION_ID
 RFDETR_PREPROCESSOR_THREADED_EXACT_V1 = "threaded-exact-v1"
 RFDETR_PREPROCESSOR_TRITON_UNIVERSAL_V1 = "triton-universal-v1"
 RFDETR_PREPROCESSOR_ENV_NAME = "INFERENCE_MODELS_RFDETR_PREPROCESSOR"
@@ -9,21 +10,6 @@ RFDETR_PREPROCESSOR_MAX_WORKERS_ENV_NAME = (
     "INFERENCE_MODELS_RFDETR_PREPROCESSOR_MAX_WORKERS"
 )
 RFDETR_PREPROCESSOR_DEFAULT_MAX_WORKERS = 4
-RFDETR_PREPROCESSOR_IDS = (
-    RFDETR_PREPROCESSOR_BASE,
-    RFDETR_PREPROCESSOR_THREADED_EXACT_V1,
-    RFDETR_PREPROCESSOR_TRITON_UNIVERSAL_V1,
-)
-
-RFDETR_POSTPROCESSOR_BASE = "base"
-RFDETR_POSTPROCESSOR_AUTO = "auto"
+RFDETR_POSTPROCESSOR_BASE = BASE_IMPLEMENTATION_ID
 RFDETR_POSTPROCESSOR_TRITON_FUSED_V1 = "triton-fused-v1"
 RFDETR_POSTPROCESSOR_ENV_NAME = "INFERENCE_MODELS_RFDETR_POSTPROCESSOR"
-RFDETR_POSTPROCESSOR_IDS = (
-    RFDETR_POSTPROCESSOR_BASE,
-    RFDETR_POSTPROCESSOR_TRITON_FUSED_V1,
-)
-
-RFDETR_BUFFER_STRATEGY_BASE = "base"
-RFDETR_SCHEDULER_BASE = "base"
-RFDETR_ENGINE_PLUGIN_BASE = "base"
