+++ inference_models/inference_models/models/rfdetr/optimization/preprocessors/triton_universal.py
@@ -2,16 +2,18 @@
 
 import torch
 
-from inference_models.models.rfdetr.optimization.contracts import (
+from inference_models.models.optimization.contracts import (
     DeviceCompatibility,
     ExecutionContext,
     InputCompatibility,
     OptimizationMetadata,
     OptimizationStage,
-    PreprocessRequest,
-    PreprocessResult,
     immutable_mapping,
     metadata_supports_context,
+)
+from inference_models.models.rfdetr.optimization.contracts import (
+    PreprocessRequest,
+    PreprocessResult,
 )
 from inference_models.models.rfdetr.optimization.ids import (
     RFDETR_PREPROCESSOR_BASE,
