+++ inference_models/tests/unit_tests/models/rfdetr/test_triton_universal_preprocess_runtime.py
@@ -3,12 +3,13 @@ import pytest
 import torch
 
 from inference_models.errors import ModelRuntimeError
-from inference_models.models.rfdetr.pre_processing import (
-    RFDETR_PREPROCESSOR_BASE,
+from inference_models.models.rfdetr.optimization.catalog import (
     RFDETR_PREPROCESSOR_IMPLEMENTATIONS,
-    RFDETR_PREPROCESSOR_THREADED_EXACT_V1,
+)
+from inference_models.models.rfdetr.optimization.ids import (
     RFDETR_PREPROCESSOR_TRITON_UNIVERSAL_V1,
-    resolve_rfdetr_preprocessor,
+)
+from inference_models.models.rfdetr.pre_processing import (
     resolve_rfdetr_preprocessor_max_workers,
 )
 from inference_models.models.rfdetr.triton_universal_preprocess_runtime import (
@@ -99,45 +100,11 @@ def test_metadata_batch_describes_stretch() -> None:
 
 
 def test_universal_candidate_is_explicitly_selectable() -> None:
-    assert (
-        resolve_rfdetr_preprocessor(RFDETR_PREPROCESSOR_TRITON_UNIVERSAL_V1)
-        == RFDETR_PREPROCESSOR_TRITON_UNIVERSAL_V1
-    )
     metadata = RFDETR_PREPROCESSOR_IMPLEMENTATIONS[
         RFDETR_PREPROCESSOR_TRITON_UNIVERSAL_V1
     ]
     assert metadata.validated_environments == ()
     assert metadata.fallback_id == "base"
-
-
-def test_preprocessor_can_be_selected_from_environment(monkeypatch) -> None:
-    monkeypatch.setenv(
-        "INFERENCE_MODELS_RFDETR_PREPROCESSOR",
-        RFDETR_PREPROCESSOR_TRITON_UNIVERSAL_V1,
-    )
-
-    assert resolve_rfdetr_preprocessor() == RFDETR_PREPROCESSOR_TRITON_UNIVERSAL_V1
-
-
-def test_preprocessor_defaults_to_base_without_environment(monkeypatch) -> None:
-    monkeypatch.delenv("INFERENCE_MODELS_RFDETR_PREPROCESSOR", raising=False)
-
-    assert resolve_rfdetr_preprocessor() == RFDETR_PREPROCESSOR_BASE
-
-
-def test_explicit_preprocessor_overrides_environment(monkeypatch) -> None:
-    monkeypatch.setenv(
-        "INFERENCE_MODELS_RFDETR_PREPROCESSOR",
-        RFDETR_PREPROCESSOR_TRITON_UNIVERSAL_V1,
-    )
-
-    assert (
-        resolve_rfdetr_preprocessor(RFDETR_PREPROCESSOR_THREADED_EXACT_V1)
-        == RFDETR_PREPROCESSOR_THREADED_EXACT_V1
-    )
-    assert resolve_rfdetr_preprocessor(RFDETR_PREPROCESSOR_BASE) == (
-        RFDETR_PREPROCESSOR_BASE
-    )
 
 
 def test_preprocessor_worker_limit_can_be_selected_from_environment(
